ReadMe
•	Briefly explain the work that you did on this project: What code were you given? What code did you create yourself?
The code we were give was about bases for the pirate game. We had to add the part where the pirate learned how to search the map itself. We used multiple AI approaches to show how they would affect the pirate in the game. Overtime, you can see the pirate learning from its mistakes and getting better and faster. It does this thanks to rewards and punishments based off each move it makes. 

•	Connect your learning from throughout this course to the larger field of computer science:
This type of AI is already being used in videogames. Several years ago, a team of computer scientist used AI to play StarCraft. They took the AI to fight the world record holder (team) and the team ended up losing to the AI. I can see this in 5 years where teams of computer scientist set up AI to fight other AI teams in videogames to see which team made the better program. 

o	What do computer scientists do and why does it matter?
Computer scientists advance the possibility of what is and isn’t possible. About twenty years about 21 years old. We have seen google take over the world in web browsing, search, phones, and even email. Just in the last twenty years this has changed everyone’s life. If google never existed, amazon might not be what it is now. Facebook might not even exist.

o	How do I approach a problem as a computer scientist?
Depending on the problem there are multiple ways to approach it. For example, you can make a list of all the steps you need, or a graph tree that shows which parts connect to each other. It all depends on the problem at hand first.

o	What are my ethical responsibilities to the end user and the organization?
As an end user, we aren’t given as much responsibility as the person who made the code. However, we can make sure we do what we think is right with the technology at hand. Just because Facebook allows you to stalk an ex doesn’t mean you should be able to. Also, any major bugs should be reported so the company can fix them before they are exploited.